fix: cannot create WASI() object with no options & default bindings not working as expected
The error was being hid by the lack of bindings on the WASI object during tests. It appears we use a build time process to include the right bindings and assign it to a static property on the object. We don't actually use the static property however when constructing the WASI object. This patch fixes that as well. And if that's fixed there doesn't seem to be a reason not to use node bindings in the tests which no longer hides the error of not being able to construct without options.