Remove the global `Buffer` dependency
The global `Buffer` is used for loading base64 encoded string.
However, this **implicit** and **global** dependency makes some
confusions and problems (#294, #305, and ruby/ruby.wasm#182).

The use of `Buffer` here is limited, so we can easily replace this
by using `atob` function. This simplifies user's build settings
and reduces the file size since there is no need to include Buffer.